refactor: 공통함수 정리
